refactor: ViewController 추가
