+++ src/main/java/roomescape/controller/ReservationController.java
@@ -5,27 +5,22 @@ import java.net.URI;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.model.domain.Reservation;
 import roomescape.model.dto.request.ReservationRequest;
 import roomescape.model.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
-@Controller
+@RestController
 @RequiredArgsConstructor
 public class ReservationController {
 
     private final ReservationService reservationService;
-
-    @GetMapping("/reservation")
-    public String Reservation() {
-        return "new-reservation";
-    }
 
     @GetMapping("/reservations")
     public ResponseEntity<List<Reservation>> getReservations() {
@@ -42,7 +37,7 @@ public class ReservationController {
 
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.deleteReservation(id);
+        reservationService.deleteReservationById(id);
         return ResponseEntity.noContent().build();
     }
 }
