+++ src/main/java/roomescape/controller/TimeController.java
@@ -5,27 +5,22 @@ import java.net.URI;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.model.domain.Time;
 import roomescape.model.dto.request.TimeRequest;
 import roomescape.model.dto.response.TimeResponse;
 import roomescape.service.TimeService;
 
-@Controller
+@RestController
 @RequiredArgsConstructor
 public class TimeController {
 
     private final TimeService timeService;
-
-    @GetMapping("/time")
-    public String Reservation() {
-        return "time";
-    }
 
     @GetMapping("/times")
     public ResponseEntity<List<Time>> getTimes() {
@@ -42,7 +37,7 @@ public class TimeController {
 
     @DeleteMapping("/times/{id}")
     public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
-        timeService.deleteTime(id);
+        timeService.deleteTimeById(id);
         return ResponseEntity.noContent().build();
     }
 }
