+++ src/main/java/roomescape/controller/ViewController.java
@@ -4,10 +4,20 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class HomeController {
+public class ViewController {
 
     @GetMapping("/")
     public String home() {
         return "index";
+    }
+
+    @GetMapping("/reservation")
+    public String Reservation() {
+        return "new-reservation";
+    }
+
+    @GetMapping("/time")
+    public String Time() {
+        return "time";
     }
 }
